add tiny-pkg-config
